Add InterpreterPromise class
